requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.2
django-autoslug==1.9.9
django-ckeditor-5==0.2.15
django-jazzmin==3.0.1
django-paypal==2.1
django-widget-tweaks==1.5.0
idna==3.10
pillow==11.0.0
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1
urllib3==2.3.0
